Basic reading and writing is working.
As is, the script will copy all of the contents of the rows containing
the first name, last name, and email address of the patron from the
Agile .xls file and to a new file called 'Email Adds.xls'.

There are various bugs in this initial commit and missing functionality.