Wrote new function that takes care of reading from multiple columns.
The original functions [firstName, lastName, emails] were only commented
out as backup.

To get the new function working, a class Column was created that, when
instantiated, sets the column the copy() function needs and sets a caps
variable to either True or False (so the copy() function can either
capitalize names or skip if emails are being pulled)

Feels pretty messy but the code currently works.