Set up initial API connection script.
cc.py can connect to the Constant Contact API and bulk upload contacts
from a JSON object. It imports from a credentials module that contains
the API key, token, and list IDs.

app.py has been moved to a separate folder (spreadsheets) to reflect
the concurrent development of two versions: the original that copies
data into a spreadsheet and a new one that hooks directly into the
Constant Contact API, as described above.