This resolves #3.
The script has three main functions:

append_dict_to_list: Called in a for loop that runs over each row in the
spreadsheet opened at the beginning. The function appends the customer
information into a dictionary contacts with one field (address) being
another dictionary; the contacts dict is then appended to either a list
for After Hours or General.

create_payload: Takes in the list from above and the corresponding ID of
the Constant Contact list that will take the contacts. It returns a
variable called payload that points to an OrderedDict that follows the
JSON structure required by the Constant Contact API.

cc: Handles the request to the Constant Contact API by taking in
create_payload as a parameter so that the OrderedDict that is returned
is set as the value of the JSON parameter in the POST request.