Created function to refresh access token.
The Gmail auth script now uses a function called gmail_auth() to
authenticate a user. It currently doesn't return the tokens.

A new refresh() function was added to the same script in order to
refresh the access token, which expires in 3600 seconds, using the
refresh token generated by gmail_auth(), which never expires.

In member.py, the code for listing a user's messages was placed inside
function list_messages(), which returns a list of all the messages
matching the parameters passed in (the label ID and the date the
messages were sent).