Create contact dictionary using email body.
The easiest way to parse the body is to use the email module to create
a string that's split.

get_contact() needs to be severely cleaned up.

The create_payload() and add_contacts() functions are nearly identical
to the ones from cc.py/cc_upload.py. Consider moving these into a
general module that can be imported.