Request membership activity as an XML file
Make GET request to URL provided by Agile that accesses their reporting
portal. The XML file in the response contains all the membership sales
activity for a given time period, in this case 'yesterday'.

Add a custom field to the column_names array in create_payload() that
will contain the membership level of the customer.