Add functions to create & send message via Gmail API per #26
These functions will confirm that an activity was completed via the
add_contacts() function in constantcontact.py. The message sent will
contain the number of errors and what they were, if any exist.

create_message() needs to be called from member.py and have parameters
passed in. This has not been tested as of this commit.