Apply while loop to call poll_activity()
Bulk activites on Constant Contact take several seconds to finish.
The while loop needs to repeatedly call poll_activity() until the
status of the report is COMPLETE or ERROR.

In the next commit, the loop should make a call to create_message() in
order to prep the email that will be sent.