Fix bug that threw AttribueError due to tuple encoding
The tuple containing the message body in member.py threw an
AttributeError when it was passed to create_message() for encoding. It
seems the error was caused by placing the str.format() method in the
wrong part.

This commit also updates the lists members are added to: the general
list in addition to the member one. As a result, the list IDs are
not wrapped in a variable within the body of create_payload() but
rather are wrapped in a list and passed to create_payload() when passed
to the function.