Turn while loop in member.py into function in constantcontact
In order to create and send emails from both member.py and cc_upload.py,
the while loop that polls the status of an activity has been turned into
a function called poll_activity() in the constantcontact module. The
function that originally had this name has been renamed get_status() as
that better reflects the function's purpose.

The subject lines and bodies of the emails generated by the
constantcontact module have been tweaked to reflect that both regular
contacts and members are added to Constant Contact.

The docstrings in constantcontact were updated.