Fix membership level not being properly uploaded
The value of the key 'name' in the membership dictionary must be
'Custom Field 1' in order to match with the column name in the payload.

Also fixed the primary member in a dual membership being overwritten by
the secondary if both entries share the same email address. Constant
Contact treats the second one as updating the first since emails
are the unique identifier for contacts.