Fix bug that caused script to fail when only one contact added
The if statement at the end of the script worked incorrectly. When
there was only one contact to add, the second conditional would return
False and prevent the contact details from being added to the members
list.

When the list was passed to the Constant Contact functions, the post
request would fail because an empty array was being passed in and the
request requires at least 1 element.

Add a list to temporarily store emails and use that to check against
new entries. This prevents the second member of a dual membership from
overriding the main member's entry.